Extract parseStackageVersions
