+++ package.yaml
@@ -66,7 +66,9 @@ tests:
     source-dirs: test
     dependencies:
       - Blammo
+      - bytestring
       - containers
+      - file-embed
       - hspec
       - hspec-expectations-lifted
       - stack-lint-extra-deps
